+++ src/main/Exception/InvalidArgumentException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace WebArch\LogTools\Exception;
+
+use InvalidArgumentException as CommonInvalidArgumentException;
+
+class InvalidArgumentException extends CommonInvalidArgumentException implements LogToolsExceptionInterface
+{
+}