+++ src/main/Exception/LogToolsExceptionInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace WebArch\LogTools\Exception;
+
+interface LogToolsExceptionInterface
+{
+}