+++ src/test/LogExceptionTraitTest.php
@@ -11,15 +11,17 @@ use WebArch\LogTools\Test\Fixture\LogExceptionFixture;
 
 class LogExceptionTraitTest extends LogExceptionFixture
 {
+    private const DUMMY_TRACE = 'trace';
+
     /**
      * @var FooService
      */
-    private $fooService;
+    private FooService $fooService;
 
     /**
      * @var TestHandler
      */
-    private $testHandler;
+    private TestHandler $testHandler;
 
     /**
      * @inheritDoc
@@ -38,202 +40,91 @@ class LogExceptionTraitTest extends LogExceptionFixture
         );
     }
 
-    public function testLogExceptionDefaultChaining()
+    public function testLogExceptionDefaultChaining(): void
     {
         $this->fooService->triggerExceptionCascade(null);
 
         $expectedRecords = [
             [
-                'message' => '[InvalidArgumentException] Some argument is invalid (1) in /log-tools/src/test/Fixture/FooService.php:*',
-                'context' => [
+                'message'    => '[InvalidArgumentException] Some argument is invalid (1) in /log-tools/src/test/Fixture/FooService.php:*',
+                'context'    => [
                     'foo'      => 'bar',
-                    'trace'    =>
-                        [
-                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
-                            1  => '#1 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(NULL)',
-                            2  => '#2 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionDefaultChaining()',
-                            3  => '#3 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
-                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
-                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
-                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
-                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
-                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
-                            11 => '#11 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
-                            12 => '#12 {main}',
-                        ],
+                    'trace'    => [self::DUMMY_TRACE],
                     'previous' => '[RuntimeException] Something is not found (2) in /log-tools/src/test/Fixture/FooService.php:*',
                 ],
-                'level_name'   => 'EMERGENCY',
+                'level_name' => 'EMERGENCY',
             ],
             [
-                'message' => '[RuntimeException] Something is not found (2) in /log-tools/src/test/Fixture/FooService.php:*',
-                'context' => [
+                'message'    => '[RuntimeException] Something is not found (2) in /log-tools/src/test/Fixture/FooService.php:*',
+                'context'    => [
                     'foo'      => 'bar',
-                    'trace'    =>
-                        [
-                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerRuntimeException()',
-                            1  => '#1 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
-                            2  => '#2 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(NULL)',
-                            3  => '#3 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionDefaultChaining()',
-                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
-                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
-                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
-                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
-                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
-                            11 => '#11 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
-                            12 => '#12 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
-                            13 => '#13 {main}',
-                        ],
+                    'trace'    => [self::DUMMY_TRACE],
                     'previous' => '[LogicException] The error in the code (3) in /log-tools/src/test/Fixture/FooService.php:*',
                 ],
-                'level_name'   => 'EMERGENCY',
+                'level_name' => 'EMERGENCY',
             ],
             [
-                'message' => '[LogicException] The error in the code (3) in /log-tools/src/test/Fixture/FooService.php:*',
-                'context' => [
+                'message'    => '[LogicException] The error in the code (3) in /log-tools/src/test/Fixture/FooService.php:*',
+                'context'    => [
                     'foo'   => 'bar',
-                    'trace' =>
-                        [
-                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerLogicException()',
-                            1  => '#1 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerRuntimeException()',
-                            2  => '#2 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
-                            3  => '#3 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(NULL)',
-                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionDefaultChaining()',
-                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
-                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
-                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
-                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
-                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            11 => '#11 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
-                            12 => '#12 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
-                            13 => '#13 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
-                            14 => '#14 {main}',
-                        ],
+                    'trace' => [self::DUMMY_TRACE],
                 ],
-                'level_name'   => 'EMERGENCY',
+                'level_name' => 'EMERGENCY',
             ],
 
         ];
         $this->assertLogRecords($expectedRecords);
     }
 
-    public function testLogExceptionChainingEnabled()
+    public function testLogExceptionChainingEnabled(): void
     {
         $this->fooService->triggerExceptionCascade(true);
 
         $expectedRecords = [
             [
-                'message' => '[InvalidArgumentException] Some argument is invalid (1) in /log-tools/src/test/Fixture/FooService.php:*',
-                'context' => [
+                'message'    => '[InvalidArgumentException] Some argument is invalid (1) in /log-tools/src/test/Fixture/FooService.php:*',
+                'context'    => [
                     'foo'      => 'bar',
-                    'trace'    =>
-                        [
-                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
-                            1  => '#1 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(true)',
-                            2  => '#2 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionChainingEnabled()',
-                            3  => '#3 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
-                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
-                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
-                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
-                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
-                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
-                            11 => '#11 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
-                            12 => '#12 {main}',
-                        ],
+                    'trace'    => [self::DUMMY_TRACE],
                     'previous' => '[RuntimeException] Something is not found (2) in /log-tools/src/test/Fixture/FooService.php:*',
                 ],
-                'level_name'   => 'EMERGENCY',
+                'level_name' => 'EMERGENCY',
             ],
             [
-                'message' => '[RuntimeException] Something is not found (2) in /log-tools/src/test/Fixture/FooService.php:*',
-                'context' => [
+                'message'    => '[RuntimeException] Something is not found (2) in /log-tools/src/test/Fixture/FooService.php:*',
+                'context'    => [
                     'foo'      => 'bar',
-                    'trace'    =>
-                        [
-                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerRuntimeException()',
-                            1  => '#1 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
-                            2  => '#2 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(true)',
-                            3  => '#3 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionChainingEnabled()',
-                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
-                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
-                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
-                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
-                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
-                            11 => '#11 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
-                            12 => '#12 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
-                            13 => '#13 {main}',
-                        ],
+                    'trace'    => [self::DUMMY_TRACE],
                     'previous' => '[LogicException] The error in the code (3) in /log-tools/src/test/Fixture/FooService.php:*',
                 ],
-                'level_name'   => 'EMERGENCY',
+                'level_name' => 'EMERGENCY',
             ],
             [
-                'message' => '[LogicException] The error in the code (3) in /log-tools/src/test/Fixture/FooService.php:*',
-                'context' => [
+                'message'    => '[LogicException] The error in the code (3) in /log-tools/src/test/Fixture/FooService.php:*',
+                'context'    => [
                     'foo'   => 'bar',
-                    'trace' =>
-                        [
-                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerLogicException()',
-                            1  => '#1 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerRuntimeException()',
-                            2  => '#2 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
-                            3  => '#3 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(true)',
-                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionChainingEnabled()',
-                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
-                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
-                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
-                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
-                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            11 => '#11 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
-                            12 => '#12 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
-                            13 => '#13 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
-                            14 => '#14 {main}',
-                        ],
+                    'trace' => [self::DUMMY_TRACE],
                 ],
-                'level_name'   => 'EMERGENCY',
+                'level_name' => 'EMERGENCY',
             ],
 
         ];
         $this->assertLogRecords($expectedRecords);
     }
 
-    public function testLogExceptionChainingDisabled()
+    public function testLogExceptionChainingDisabled(): void
     {
         $this->fooService->triggerExceptionCascade(false);
 
         $expectedRecords = [
             [
-                'message' => '[InvalidArgumentException] Some argument is invalid (1) in /log-tools/src/test/Fixture/FooService.php:*',
-                'context' => [
+                'message'    => '[InvalidArgumentException] Some argument is invalid (1) in /log-tools/src/test/Fixture/FooService.php:*',
+                'context'    => [
                     'foo'      => 'bar',
-                    'trace'    =>
-                        [
-                            0  => '#0 /log-tools/src/test/Fixture/FooService.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerInvalidArgumentException()',
-                            1  => '#1 /log-tools/src/test/LogExceptionTraitTest.php(*): WebArch\\LogTools\\Test\\Fixture\\FooService->triggerExceptionCascade(false)',
-                            2  => '#2 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): WebArch\\LogTools\\Test\\LogExceptionTraitTest->testLogExceptionChainingDisabled()',
-                            3  => '#3 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestCase->runTest()',
-                            4  => '#4 /log-tools/vendor/phpunit/phpunit/src/Framework/TestResult.php(*): PHPUnit\\Framework\\TestCase->runBare()',
-                            5  => '#5 /log-tools/vendor/phpunit/phpunit/src/Framework/TestCase.php(*): PHPUnit\\Framework\\TestResult->run(Object(WebArch\\LogTools\\Test\\LogExceptionTraitTest))',
-                            6  => '#6 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestCase->run(Object(PHPUnit\\Framework\\TestResult))',
-                            7  => '#7 /log-tools/vendor/phpunit/phpunit/src/Framework/TestSuite.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            8  => '#8 /log-tools/vendor/phpunit/phpunit/src/TextUI/TestRunner.php(*): PHPUnit\\Framework\\TestSuite->run(Object(PHPUnit\\Framework\\TestResult))',
-                            9  => '#9 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\TestRunner->doRun(Object(PHPUnit\\Framework\\TestSuite), Array, Array, true)',
-                            10 => '#10 /log-tools/vendor/phpunit/phpunit/src/TextUI/Command.php(*): PHPUnit\\TextUI\\Command->run(Array, true)',
-                            11 => '#11 /log-tools/vendor/phpunit/phpunit/phpunit(61): PHPUnit\\TextUI\\Command::main()',
-                            12 => '#12 {main}',
-                        ],
+                    'trace'    => [self::DUMMY_TRACE],
                     'previous' => '[RuntimeException] Something is not found (2) in /log-tools/src/test/Fixture/FooService.php:*',
                 ],
-                'level_name'   => 'EMERGENCY',
+                'level_name' => 'EMERGENCY',
             ],
         ];
         $this->assertLogRecords($expectedRecords);
@@ -258,6 +149,10 @@ class LogExceptionTraitTest extends LogExceptionFixture
                 $record['level_name'],
                 'Equal level for record #' . $key
             );
+            // Simplify 'trace' since it depends on the way the test is run
+            if (key_exists('trace', $record['context'])) {
+                $record['context']['trace'] = [self::DUMMY_TRACE];
+            }
             $this->assertEqualsCanonicalizing(
                 $expectedRecord['context'],
                 $this->cleanUpArray($record['context']),